refactor(settings): move SettingsService to sub-namespace
SettingsService has been moved from CastorApplication.Services
to the new CastorApplication.Services.Settings namespace for
better code organization and clarity.
